Add Makefile and .helmignore files for kibana, filebeat, logstash, and elasticsearch
This commit adds the Makefile and .helmignore files for the kibana, filebeat, logstash, and elasticsearch directories. These files are necessary for building and packaging the respective components. The Makefile includes the common.mk file from the helpers directory, while the .helmignore file specifies the directories to be ignored when packaging the helm charts.